refactor: move function to search.py
